backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
sqlalchemy
pymysql
python-dotenv